Restructure events
